Extract embedded Asciidoctor samples discovery to new library
Embedded Asciidoctor sources discovery requires transitive
dependencies on JRuby and a whole lot of other libraries.

This change creates a new samples-discovery-asciidoctor library
that one can put on their testImplementation configuration separately
thus allowing users to opt in.
