+++ sample-discovery-asciidoctor/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    groovy
+}
+
+dependencies {
+    compileOnly(Libraries.JSR305)
+    implementation(project(":sample-discovery"))
+    implementation(Libraries.ASCIIDOCTORJ)
+    implementation(Libraries.ASCIIDOCTORJ_API)
+    implementation(Libraries.COMMONS_IO)
+    implementation(Libraries.COMMONS_LANG3)
+    implementation(Libraries.TYPESAFE_CONFIG)
+    testImplementation(Libraries.SPOCK_CORE)
+}
