+++ sample-discovery-asciidoctor/src/test/groovy/org/gradle/samples/loader/asciidoctor/AsciidoctorSamplesDiscoveryTest.groovy
