+++ settings.gradle.kts
@@ -1,6 +1,7 @@
 rootProject.name = "exemplar"
 
 include("sample-discovery")
+include("sample-discovery-asciidoctor")
 include("sample-check")
 
 enableFeaturePreview("STABLE_PUBLISHING")
