+++ src/test/java/org/gradle/exemplar/GradleEmbeddedSamplesRunner.java
@@ -13,34 +13,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.samples.test.runner;
+package org.gradle.exemplar;
 
-import org.gradle.samples.loader.SamplesDiscovery;
+import org.gradle.samples.loader.EmbeddedSamplesDiscovery;
 import org.gradle.samples.model.Sample;
+import org.gradle.samples.test.runner.GradleSamplesRunner;
+import org.gradle.samples.test.runner.SamplesRunner;
 import org.junit.runners.model.InitializationError;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
-public class EmbeddedSamplesRunner extends SamplesRunner {
-
+/**
+ * A custom implementation of {@link SamplesRunner} that uses the Gradle Tooling API to execute sample builds.
+ */
+public class GradleEmbeddedSamplesRunner extends GradleSamplesRunner {
     /**
-     * Constructs a new {@code ParentRunner} that will run {@code @TestClass}
-     *
-     * @param testClass reference to test class being run
+     * {@inheritDoc}
      */
-    public EmbeddedSamplesRunner(Class<?> testClass) throws InitializationError {
+    public GradleEmbeddedSamplesRunner(Class<?> testClass) throws InitializationError {
         super(testClass);
     }
 
     @Override
     protected List<Sample> getChildren() {
-        File samplesRootDir = getSamplesRootDir();
         try {
-            return SamplesDiscovery.embeddedSamples(samplesRootDir);
+            return EmbeddedSamplesDiscovery.embeddedSamples(getSamplesRootDir());
         } catch (IOException e) {
-            throw new RuntimeException("Could not extract samples from " + samplesRootDir, e);
+            throw new RuntimeException("Could not extract embedded samples", e);
         }
     }
 }
